style: reuse collapsible card in visualization card, remove unnecesary code
