+++ packages/frontend/src/components/Explorer/Explorer.styles.ts
@@ -1,4 +1,3 @@
-import { Card, Collapse } from '@blueprintjs/core';
 import styled from 'styled-components';
 
 export const ExplorerWrapper = styled.div`
@@ -6,25 +5,4 @@ export const ExplorerWrapper = styled.div`
     flex-direction: column;
     gap: 10px;
     flex: 1;
-`;
-
-interface StyledCollapseProps {
-    isOpen?: boolean;
-    $flexGrow?: number;
-}
-
-export const StyledCollapse = styled(Collapse)<StyledCollapseProps>`
-    flex-grow: ${({ $flexGrow }) => $flexGrow || 'unset'};
-    min-height: ${({ isOpen }) => (isOpen ? '300px' : 'unset')};
-
-    .bp4-collapse-body {
-        height: 100%;
-    }
-`;
-
-export const ExploreCard = styled(Card)<{ $flexGrow?: number }>`
-    flex-grow: ${({ $flexGrow }) => $flexGrow || 'unset'};
-    padding: 5px;
-    display: flex;
-    flex-direction: column;
 `;
